Add scripts provision clusters with STS
This commits adds a boolean variable to enable
STS that is false by default for backward compatibility
of this module.

The calls to the rosa CLI had to be added to the script that
creates and deletes the cluster instead of a new resource in
terraform because the STS roles and oidc provider can only be
deleted using the cluster ID that is only known after the
cluster is created.

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>